+++ pikachu_education/lib/bloc/bloc_login_page/login_bloc.dart
@@ -2,7 +2,7 @@ import 'package:bloc/bloc.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:meta/meta.dart';
 import 'package:pikachu_education/service/service_local_storage/service_save_data_to_local_storage.dart';
-import 'package:pikachu_education/service/service_login/firebase_login.dart';
+import 'package:pikachu_education/service/service_login/firebase_login_by_google.dart';
 import 'package:pikachu_education/service/service_login/firebase_login_by_phone_number.dart';
