+++ pikachu_education/lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
-import 'package:pikachu_education/pages/authentication/login_page/verify_otp_page.dart';
 import 'package:pikachu_education/pages/authentication/login_page/login_page.dart';
 import 'package:pikachu_education/routes/route_management.dart';
 import 'package:firebase_core/firebase_core.dart';
@@ -25,7 +24,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: LoginPage(),
+      home: const LoginPage(),
       localizationsDelegates: const [FormBuilderLocalizations.delegate],
       onGenerateRoute: generateRoute,
     );
