+++ pikachu_education/lib/pages/authentication/login_page/login_page.dart
@@ -1,20 +1,17 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:google_sign_in/google_sign_in.dart';
 import 'package:pikachu_education/bloc/bloc_login_page/login_bloc.dart';
 import 'package:pikachu_education/components/positive_button.dart';
 import 'package:pikachu_education/components/snack_bar_custom.dart';
 import 'package:pikachu_education/components/text_form_field.dart';
 import 'package:pikachu_education/routes/page_name.dart';
-
-import 'package:pikachu_education/service/service_login/firebase_login_by_phone_number.dart';
 import 'package:pikachu_education/utils/management_image.dart';
+import 'package:pikachu_education/utils/management_regex.dart';
 
-import 'package:pikachu_education/utils/management_text.dart';
+import 'package:pikachu_education/utils/management_text_style.dart';
 
-import 'method_login.dart';
-import 'method_login_loading.dart';
+import 'component/method_login.dart';
+import 'component/method_login_loading.dart';
 
 class LoginPage extends StatefulWidget {
   const LoginPage({super.key});
@@ -25,31 +22,16 @@ class LoginPage extends StatefulWidget {
 
 class _LoginPageState extends State<LoginPage> {
   final keyOfLogin = GlobalKey<FormState>();
-
-  // final phoneController = TextEditingController();
-  // final passwordController = TextEditingController();
-  bool showPassword = true;
-  bool showPasswordIcon = true;
+  
   final LoginBloc loginBloc = LoginBloc();
   final phoneNumberController = TextEditingController();
 
   @override
   void dispose() {
     phoneNumberController.dispose();
-    // phoneController.dispose();
-    // passwordController.dispose();
     super.dispose();
   }
-
-  // Future<void> loadDataForLogin() async {
-  //   final SharedPreferences prefs = await SharedPreferences.getInstance();
-  //   var user = prefs.getString(ManagementKey.user) ?? '';
-  //   var password = prefs.getString(ManagementKey.password) ?? '';
-  //   setState(() {
-  //     phoneController.text = user;
-  //     passwordController.text = password;
-  //   });
-  // }
+  
 
   @override
   void initState() {
@@ -105,62 +87,16 @@ class _LoginPageState extends State<LoginPage> {
                               const SizedBox(height: 100),
                               Image.asset(ManagementImage.logo),
                               const SizedBox(height: 60),
-                              // TextFormFieldCustom(
-                              //     hintText: 'phone',
-                              //     textEditingController: phoneController,
-                              //     validator: (value) {
-                              //       if (value == null || value.isEmpty) {
-                              //         return 'email can not be empty';
-                              //       }
-                              //       RegExp userExp = RegExp(
-                              //           r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$');
-                              //       if (!userExp!.hasMatch(value)) {
-                              //         return 'Your User is invalid';
-                              //       }
-                              //     },
-                              //     textInputType: TextInputType.text),
-                              // const SizedBox(
-                              //   height: 20,
-                              // ),
-                              // TextFormFieldCustom(
-                              //   textEditingController: passwordController,
-                              //   validator: (value) {
-                              //     if (value == null || value.isEmpty) {
-                              //       return 'Password can not be empty';
-                              //     }
-                              //     RegExp passwordExp = RegExp(
-                              //         '^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@\$%^&*-]).{8,18}\$');
-                              //     if (!passwordExp.hasMatch(value)) {
-                              //       return 'Your Password is invalid';
-                              //     }
-                              //   },
-                              //   hintText: 'Password',
-                              //   textInputType: TextInputType.text,
-                              //   obscuringCharacter: '*',
-                              //   turnObscuringCharacter: showPasswordIcon,
-                              //   suffixIcon: IconButton(
-                              //       onPressed: () {
-                              //         setState(() {
-                              //           showPassword = !showPassword;
-                              //           showPasswordIcon = !showPasswordIcon;
-                              //         });
-                              //       },
-                              //       icon: Icon(showPasswordIcon
-                              //           ? Icons.visibility_off
-                              //           : Icons.visibility)),
-                              // ),
-                              // const SizedBox(
-                              //   height: 20,
-                              // ),
+                              
                               TextFormFieldCustom(
                                   textEditingController: phoneNumberController,
                                   validator: (value) {
                                     if (value == null || value.isEmpty) {
                                       return 'Password can not be empty';
                                     }
-                                    RegExp passwordExp =
-                                        RegExp('^[+]?[0-9]{10,15}\$');
-                                    if (!passwordExp.hasMatch(value)) {
+                                    RegExp phoneNumExp =
+                                        ManagementRegex.phoneNumber;
+                                    if (!phoneNumExp.hasMatch(value)) {
                                       return 'Your Phone Number is invalid';
                                     }
                                   },
@@ -169,27 +105,6 @@ class _LoginPageState extends State<LoginPage> {
                               const SizedBox(
                                 height: 40,
                               ),
-                              // Padding(
-                              //   padding:
-                              //   const EdgeInsets.symmetric(horizontal: 20),
-                              //   child: Row(
-                              //     mainAxisAlignment: MainAxisAlignment.end,
-                              //     children: [
-                              //       InkWell(
-                              //           onTap: () {
-                              //             Navigator.pushNamed(
-                              //                 context, PageName.getOtpPage);
-                              //           },
-                              //           child: const Text(
-                              //             'Forgot password?',
-                              //             style: TextStyle(
-                              //                 fontSize: 15,
-                              //                 decoration:
-                              //                 TextDecoration.underline),
-                              //           ))
-                              //     ],
-                              //   ),
-                              // ),
                                PositiveButtonCustom(stateLoading: state is LoginWithPhoneNumLoadingState?true:false,
                                       nameButton: 'LOGIN',
                                       onPressed: () {
@@ -210,31 +125,11 @@ class _LoginPageState extends State<LoginPage> {
                               const SizedBox(
                                 height: 20,
                               ),
-                              // PositiveButtonCustom(
-                              //     nameButton: 'LOGIN',
-                              //     onPressed: () {
-                              //       {
-                              //         keyOfLogin.currentState!.validate();
-                              //         if (keyOfLogin.currentState!.validate() ==
-                              //             true) {
-                              //           SaveDataToLocal.saveDataForLogin(
-                              //               context,
-                              //               phoneController.text,
-                              //               passwordController.text);
-                              //           context.read<LoginBloc>().add(
-                              //               LoginPressEvent(
-                              //                   email: phoneController.text,
-                              //                   password:
-                              //                       passwordController.text,
-                              //                   context: context));
-                              //         }
-                              //       }
-                              //     }),
                               const SizedBox(
                                 height: 20,
                               ),
                               Text('Or Login with',
-                                  style: ManagementText.normalStyle),
+                                  style: ManagementTextStyle.normalStyle),
                               const SizedBox(
                                 height: 20,
                               ),
