+++ pikachu_education/lib/utils/management_text_inform.dart
@@ -1,0 +1,3 @@
+class ManagementTextInform{
+  static String informOTP = 'Please enter your email and press GET OTP button, an OPT code will be send your email';
+}
