+++ pikachu_education/lib/utils/management_text_style.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class ManagementText{
+class ManagementTextStyle{
   static TextStyle normalStyle = const TextStyle(fontWeight: FontWeight.w400,fontSize: 15,color: Colors.black);
   static TextStyle methodLogin = const TextStyle(fontWeight: FontWeight.bold,fontSize: 18,color: Colors.black);
 }
